+++ plugins/flusher/kafkav2/scram_client.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package kafka
+package kafkav2
 
 import (
 	"crypto/sha256"
